requirements.txt
# Video Processing and Editing
moviepy==1.0.3
opencv-python==4.8.1.78

# AI and Transcription
openai-whisper==20231117
torch

# Input Handling
pytube==15.0.0
requests==2.31.0

# Report Generation
fpdf==1.7.2

# System Utilities
numpy